fix: improve diagram rename
