+++ packages/compass-data-modeling/src/utils/utils.ts
@@ -5,6 +5,7 @@ import type {
   Relationship,
 } from '../services/data-model-storage';
 import { cloneDeepWith } from 'lodash';
+import toNS from 'mongodb-ns';
 
 export const isIdField = (fieldPath: FieldPath): boolean =>
   fieldPath.length === 1 && fieldPath[0] === '_id';
@@ -142,4 +143,25 @@ export function mapFieldDataToJsonSchema(
     return value;
   });
   return newFieldData;
+}
+
+export function getDefaultRelationshipName(
+  relationship: Relationship['relationship']
+): string {
+  const [local, foreign] = relationship;
+  let localLabel = '';
+  let foreignLabel = '';
+  if (local.ns) {
+    localLabel += toNS(local.ns).collection;
+    if (local.fields && local.fields.length) {
+      localLabel += `.${local.fields.join('.')}`;
+    }
+  }
+  if (foreign.ns) {
+    foreignLabel += toNS(foreign.ns).collection;
+    if (foreign.fields && foreign.fields.length) {
+      foreignLabel += `.${foreign.fields.join('.')}`;
+    }
+  }
+  return [localLabel, foreignLabel].join(` \u2192 `).trim();
 }
